added modal window to change posts

+++ src/components/Post/Post.tsx
@@ -2,9 +2,13 @@ import { IPost } from "../../types/types"
 import dislike from "../../assets/like.svg"
 import like from "../../assets/likePressed.svg"
 import comment from "../../assets/comment.svg"
-import trash from '../../assets/trash.svg'
+import trash from "../../assets/trash.svg"
+import pen from "../../assets/pen.svg"
 import styles from "./post.module.scss"
-import { useGetUserByIdQuery, useGetUserQuery } from "../../store/Api/userApislice"
+import {
+    useGetUserByIdQuery,
+    useGetUserQuery,
+} from "../../store/Api/userApislice"
 import {
     useCheckLikeQuery,
     useDeletePostMutation,
@@ -12,7 +16,9 @@ import {
 } from "../../store/Api/postsSlice"
 import { FC } from "react"
 import { Link, useNavigate } from "react-router-dom"
-import { useAppSelector } from "../../hooks/redux"
+import { useAppDispatch, useAppSelector } from "../../hooks/redux"
+import UpdatePostModal from "../UpdatePostModal/UpdatePostModal"
+import { setUpdatePostDto, toggleUpdateModal } from "../../store/Slices/InterfaceSlice"
 
 export interface PostProps {
     post: IPost
@@ -23,20 +29,20 @@ export const Post: FC<PostProps> = ({ post }) => {
     const router = useNavigate()
     const [likePost] = useLikePostMutation()
     const [deletePost] = useDeletePostMutation()
-    const { data: user } = useGetUserByIdQuery(post.userId)    
-    
+    const { data: user } = useGetUserByIdQuery(post.userId)
+
     const { data: isLiked } = useCheckLikeQuery(post.postId, {
         skip: !isLogged,
     })
-
-    const {data: currentUser} = useGetUserQuery(undefined)
+    const { data: currentUser } = useGetUserQuery(undefined)
+    const dispatch = useAppDispatch()
 
     async function handleLikePost(likeid: number) {
         try {
             if (isLogged) {
                 const res = await likePost(likeid)
             } else {
-                router('/login')
+                router("/login")
             }
         } catch (e) {
             console.log(e)
@@ -45,12 +51,10 @@ export const Post: FC<PostProps> = ({ post }) => {
 
     async function handleDeletePost(postId: number) {
         try {
-            if (confirm('delete post?'))
-            await deletePost(postId)
-            router('/')
+            if (confirm("delete post?")) await deletePost(postId)
+            router("/")
         } catch (e) {
-            console.log(e);
-            
+            console.log(e)
         }
     }
 
@@ -60,7 +64,9 @@ export const Post: FC<PostProps> = ({ post }) => {
                 <span></span>
                 <div className={styles.postTitle}>
                     <h1>{post.title}</h1>
-                    <Link to={`/users/${post.userId}/posts`}>@{user?.userName}</Link>
+                    <Link to={`/users/${post.userId}/posts`}>
+                        @{user?.userName}
+                    </Link>
                 </div>
             </div>
             <div className={styles.postPlot}>{post.plot}</div>
@@ -82,14 +88,31 @@ export const Post: FC<PostProps> = ({ post }) => {
                         onClick={() => router(`/comments/${post.postId}`)}
                     />
                 </div>
-                {currentUser?.userName === user?.userName && <div className={styles.buttonSection}>
-                    <img
-                        src={trash}
-                        width="24"
-                        alt="comment"
-                        onClick={() => {handleDeletePost(post.postId)}}
-                    />
-                </div>}
+                {currentUser?.userName === user?.userName && (
+                    <>
+                        <div className={styles.buttonSection}>
+                            <img
+                                src={trash}
+                                width="24"
+                                alt="comment"
+                                onClick={() => {
+                                    handleDeletePost(post.postId)
+                                }}
+                            />
+                        </div>
+                        <div className={styles.buttonSection}>
+                            <img
+                                src={pen}
+                                width="24"
+                                alt="comment"
+                                onClick={() => {
+                                    dispatch(setUpdatePostDto({updatePostDto: post}))
+                                    dispatch(toggleUpdateModal())
+                                }}
+                            />
+                        </div>
+                    </>
+                )}
             </div>
         </div>
     )
